🔧 more trials! up is never far away. use new psiclj
double habit phase trials 8*6->16*6
up is perceptually different than left or right. so up is never far/best
using new not-found.html that needs /tasks. so update psiclj w/-t parameter

this is a bit of a hack. there seems to be a bug in native-image
version of psiclj. otherwise we could ditch our local not-found.html

and because we still have it. we don't actually need the '-t habit'
or ajax call to /tasks. we could just hardcode 'habit' into the not-found